Add support for mapping List<T> of nested types
If a field or setter is annotated with `@XPathRoot` and its type is
`List<T>`, where `T` is a type with mapping annotations of its own,
the `XPathMapper` will now create instances of `T` from every element
matching the paths in `XPathRoot`, using the sub-tree (with the matching
element as the root) for the mapping.

The unit test has two example mappings that can be refered to for
demonstration purposes.