commons-file: Add new FileSystemResourceIOException...
...to differentiate between "no readable resources found for identifier"
and "potential resource found, but I/O error prevented us from
completing the operation". We want to differentiate downstream between
the two (e.g. in a webservice the former would be a 404 and the latter a
500, raw disk i/o is definitely an internal server error).

We opted for subclassing the existing `ResourceIOException` as to not
break the existing API.